move connection status code
